Rename test section
